requirements.txt
openai-agents>=0.1.0
openai>=1.0.0
faiss-cpu>=1.7.0
numpy>=1.24.0
pyarabic>=0.6.15
python-dotenv>=1.0.0
tiktoken>=0.5.0

